% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/execute_api_request.R
\name{execute_api_request}
\alias{execute_api_request}
\title{Execute API request and return parsed content}
\usage{
execute_api_request(req_url)
}
\arguments{
\item{req_url}{URL for API request}
}
\value{
List of parsed response
}
\description{
Execute API request and return parsed content
}
