% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_data.R
\name{get_latest_data}
\alias{get_latest_data}
\title{Get latest data from weather stations}
\usage{
get_latest_data(wh_state = "CO", token = Sys.getenv("SYNOPTIC_API_KEY"))
}
\arguments{
\item{wh_state}{State to get data for}

\item{token}{token (api key). Default is Sys.getenv("SYNOPTIC_API_KEY")}
}
\value{
Dataframe
}
\description{
Get latest data from weather stations
}
