% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_metadata.R
\name{get_metadata}
\alias{get_metadata}
\title{Get station metadata from synopticdata API}
\usage{
get_metadata(wh_state = "CO", token = Sys.getenv("SYNOPTIC_API_KEY"))
}
\arguments{
\item{wh_state}{State to get info for. Default "CO"}

\item{token}{token (api key). Default is Sys.getenv("SYNOPTIC_API_KEY")}
}
\value{
df: Data frame of requested station metadata
}
\description{
Get station metadata from synopticdata API
}
\examples{
meta <- get_metadata()
}
