% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_networks.R
\name{get_networks}
\alias{get_networks}
\title{Get info on available networks}
\usage{
get_networks(token = Sys.getenv("SYNOPTIC_API_KEY"))
}
\arguments{
\item{token}{token (api key). Default is Sys.getenv("SYNOPTIC_API_KEY")}
}
\value{
Dataframe of networks and information
}
\description{
Get info on available networks
}
\examples{
networks <- get_networks()
}
