% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_timeseries_data.R
\name{get_timeseries_data}
\alias{get_timeseries_data}
\title{Get timeseries data for a weather station}
\usage{
get_timeseries_data(
  start_time = "202407100000",
  end_time = "202407150000",
  units = c("english", "metric"),
  wh_station = "KDEN",
  token = Sys.getenv("SYNOPTIC_API_KEY")
)
}
\arguments{
\item{start_time}{Starting time in YYYYmmddHHMM format.}

\item{end_time}{End time in YYYYmmddHHMM format}

\item{units}{c("english", "metric") - Default "english"}

\item{wh_station}{Station id}

\item{token}{Default is Sys.getenv("SYNOPTIC_API_KEY")}
}
\value{
Dataframe of timeseries data
}
\description{
Get timeseries data for a weather station
}
