% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/synopticr-package.R
\docType{package}
\name{synopticr-package}
\alias{synopticr}
\alias{synopticr-package}
\title{synopticr: R API Wrapper for synopticdata weather API}
\description{
Get weather data from synopticdata API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/andypicke/synopticr}
  \item Report bugs at \url{https://github.com/andypicke/synopticr/issues}
}

}
\author{
\strong{Maintainer}: Andy Pickering \email{andypicke@gmail.com} (\href{https://orcid.org/0000-0002-5724-7042}{ORCID})

}
\keyword{internal}
